Restructure config code
