+++ src/config/index.ts
@@ -1,0 +1,3 @@
+
+export * from './spawn'
+export * from './gen'
